Updating the app to use the slack user id to send messages
This is initial work to move from sending slack messages from using
the slack username to using the user id for that slack account. Ideally this
will solve the issue of when the user's slackname changes that messages are still
properly sent. It doesn't fully do this yet, but that's the end goal.

Some noteworthy changes:
- When employees are created it saves the user id to the DB
- Test message sending will update the channel and user id based on the slackname
if it can
- A variety of other class methods to help get the user id

Lots of work to be done, but it's a start