Updating to only send messages when either slack room or user id is present
This adds code to only send messages if either the slack room id
or the slack user id is present. Otherwise it either alerts to a failed message
on the test send, or it will add a sent scheduled message with an error
listed.

This should help identify when slack usernames have been changed